backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# DICOM Processing
pydicom==2.4.4
SimpleITK==2.3.1
numpy==1.26.3
Pillow==10.2.0

# AI/ML
torch==2.1.2
transformers==4.37.2
accelerate==0.26.1
bitsandbytes==0.42.0

# Medical Imaging
nibabel==5.2.0
scikit-image==0.22.0

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Integrations
hl7==0.4.5
fhir.resources==7.1.0
aiofiles==23.2.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Storage
boto3==1.34.34
azure-storage-blob==12.19.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
